Рефакторинг компонентов. Разбивка на частные подкомпоненты
